Allow user to decide if links are updated when renaming a note
By default, when renaming a note to which other notes have links,
display a dialog explaining the situation, showing a list of affected
notes, and letting the user pick which notes (if any) will be updated.
Notes that are not updated will have the link removed (not just
broken).

The dialog also allows the user to set a preference to always show the
dialog, or always update link text to reference the new note name, or
always leave the text alone and remove the link. This preference has
also been added to the Editing tab of the Preferences dialog.

Fixes: https://bugzilla.gnome.org/584789