Search window context menus rework
* Use only one context menu in notebooks list, because the second
is not always available (when there are a lot of notebooks, you
can not click on empty place, also, it is not possible to press
the context menu key on empty place). "New notebook" item has been
added to the first context menu, while the second is removed.
* Added "new note" item to notes list context menu.

Fixes bug 601852.