Optimize note creation, renaming and deletion
With a large number of notes, Gnote becomes very slow at creating,
renaming and deleting notes.

First, there is no need to throw away the whole trie and re-create it
from scratch when a new note is being created. The title of the new
note can be added to the existing trie.

Secondly there is no need to throw away and re-create the
Gtk::ListStore within the recent changes window from scratch when a
note has been created, renamed or deleted. Instead, the existing model
can be edited.

Fixes: https://bugzilla.gnome.org/660663