preferencesdialog: Fix setting margin
Setting margin on the dialog causes a 10 pixel transparent area at the
right and bottom of the window. Set spacing on the content area has no
effect as it has only one child (the notebook). Instead of that, set margin
on the parent of the content area, which gives the expected result.